Rename WebApi Project and namespace to Api
