+++ account-management/Api/Tenants/Contracts/UpdateTenantRequest.cs
@@ -1,3 +1,3 @@
-namespace PlatformPlatform.AccountManagement.WebApi.Tenants.Contracts;
+namespace PlatformPlatform.AccountManagement.Api.Tenants.Contracts;
 
 public sealed record UpdateTenantRequest(string Name, string Email, string? Phone);
