+++ account-management/Tests/ArchitectureTests/DependencyTests.cs
@@ -1,9 +1,9 @@
 using FluentAssertions;
 using NetArchTest.Rules;
+using PlatformPlatform.AccountManagement.Api;
 using PlatformPlatform.AccountManagement.Application;
 using PlatformPlatform.AccountManagement.Domain;
 using PlatformPlatform.AccountManagement.Infrastructure;
-using PlatformPlatform.AccountManagement.WebApi;
 using Xunit;
 
 namespace PlatformPlatform.AccountManagement.Tests.ArchitectureTests;
