+++ account-management/Tests/Tests.csproj
@@ -15,7 +15,7 @@
         <ProjectReference Include="..\Domain\Domain.csproj"/>
         <ProjectReference Include="..\Application\Application.csproj"/>
         <ProjectReference Include="..\Infrastructure\Infrastructure.csproj"/>
-        <ProjectReference Include="..\WebApi\WebApi.csproj"/>
+        <ProjectReference Include="..\Api\Api.csproj"/>
     </ItemGroup>
 
     <ItemGroup>
